fix(files): moved bus stop tile + small styling changes + linting
