+++ src/components/BusStop/BusStopTile/BusStopTile.module.scss
@@ -1,6 +1,4 @@
-.BusStopTile {
-    max-height: 80vh;
-}
+
 
 .TileHeader {
     padding-bottom: 2rem;
