+++ src/dashboards/BusStop/BusStopDashboard.module.scss
@@ -1,3 +1,7 @@
 .BusStopTileWrapper {
     margin-bottom: 2rem;
 }
+
+.BusStopTile {
+    max-height: 80vh;
+ }
